Created the welcome feature and the model.
The welcome feature is comprised of a very simple activity and fragment.

The model is the start of an attempt to move the architecture to a more traditional model-view-controller pattern. The model-view-viewmodel, while a good pattern on paper, has been poorly implemented and will eventually be refactored out of this project.